Move constants files to their own module
Minor cleanup in main.py
