+++ duckbot/cogs/announce_day.py
@@ -1,6 +1,6 @@
 import random
-import cogs.channels as channels
-import cogs.emojis as emojis
+import server.channels as channels
+import server.emojis as emojis
 import datetime
 import pytz
 import holidays
